add solve_step

+++ rubiks_cube.hpp
@@ -3,7 +3,7 @@
 #include <random>
 
 // U, L, F, R, B, D
-typedef std::array<int, 54> puzzle;
+typedef std::array<size_t, 54> puzzle;
 std::random_device rd;
 
 enum class Action { R, L, U, D, F, B };
@@ -22,16 +22,13 @@ puzzle act( const puzzle& cube, Action a ) {
     puzzle ret = cube;
     switch (a) {
         case Action::R :
-            // R -> R
             ret = permute5( ret, 28, 30, 36, 34, 28 );
             ret = permute5( ret, 29, 33, 35, 31, 29 );
-            // U -> B
             ret = permute5( ret, 3, 43, 52, 21, 3 );
             ret = permute5( ret, 6, 40, 49, 24, 6 );
             ret = permute5( ret, 9, 37, 46, 27, 9 );
             return ret;
         case Action::L :
-            // L -> L
             ret = permute5( ret, 10, 12, 18, 16, 10 );
             ret = permute5( ret, 11, 15, 17, 13, 11 );
             ret = permute5( ret, 1, 19, 54, 45, 1 );
@@ -39,7 +36,6 @@ puzzle act( const puzzle& cube, Action a ) {
             ret = permute5( ret, 7, 25, 48, 39, 7 );
             return ret;
         case Action::U :
-            // U -> U
             ret = permute5( ret, 1, 3, 9, 7, 1 );
             ret = permute5( ret, 2, 6, 8, 4, 2 );
             ret = permute5( ret, 19, 10, 37, 28, 19 );
@@ -47,7 +43,6 @@ puzzle act( const puzzle& cube, Action a ) {
             ret = permute5( ret, 21, 12, 39, 30, 21 );
             return ret;
         case Action::D :
-            // D -> D
             ret = permute5( ret, 46,  48,  54,  52,  46 );
             ret = permute5( ret, 47,  51,  53,  49,  47 );
             ret = permute5( ret, 25,  30,  43,  16,  25 );
@@ -55,7 +50,6 @@ puzzle act( const puzzle& cube, Action a ) {
             ret = permute5( ret, 27,  36,  45,  18,  27 );
             return ret;
         case Action::F :
-            // F -> F
             ret = permute5( ret, 19,  21,  27,  25,  19 );
             ret = permute5( ret, 20,  24,  26,  22,  20 );
             ret = permute5( ret, 7,  28,  52,  18,  7 );
@@ -63,7 +57,6 @@ puzzle act( const puzzle& cube, Action a ) {
             ret = permute5( ret, 9,  34,  54,  12,  9 );
             return ret;
         default :
-            // B ->  B
             ret = permute5( ret, 37,  39,  45,  43,  37 );
             ret = permute5( ret, 38,  42,  44,  40,  38 );
             ret = permute5( ret, 3,  10,  48,  36,  3 );
@@ -112,6 +105,5 @@ OS & operator << ( OS & os, const puzzle & state ) {
 int main() {
     puzzle state = random_puzzle();
     std::cout << goal_state;
-    //std::cout << permute5( goal_state, 1, 3, 9, 7, 1);
     std::cout << act(goal_state, Action::U);
 }
